refactor: move shared message functions to shared/messages
