+++ src/runtime/server/type-generation.ts
@@ -1,7 +1,7 @@
 import { deepCopy, isArray, isFunction, isObject } from '@intlify/shared'
 import { vueI18nConfigs, localeLoaders, normalizedLocales } from '#internal/i18n/options.mjs'
 import { dtsFile } from '#internal/i18n-type-generation-options'
-import { loadVueI18nOptions } from '../messages'
+import { loadVueI18nOptions } from '../shared/messages'
 import { getMergedMessages } from './utils/messages'
 import { writeFile } from 'fs/promises'
 import { useRuntimeConfig } from '#imports'
